\name{SMI}
\alias{SMI}
\alias{y}
\docType{data}
\title{Swiss Market Index data}
\usage{
data("SMI")
}
\description{
  See Mullen et al. (2011) for description of this dataset.
}
\references{
  Mullen, K.M, Ardia, D., Gil, D., Windover, D., Cline,J. (2011). 
  \pkg{DEoptim:} An R Package for Global Optimization by Differential Evolution. 
  \emph{Journal of Statistical Software}, 40(6), 1-26.
  \doi{10.18637/jss.v040.i06}
}
\keyword{datasets}
