\name{xrrData}
\alias{xrrData}

\docType{data}
\title{Example X-ray reflectometry data}
\usage{
data("xrrData")
}
\description{See
  Mullen, K.M., Ardia, D., Gil, D.L, Windover, D., Cline, J. (2009).
  DEoptim: An R Package for Global Optimization by Differential Evolution.
  \url{http://ssrn.com/abstract=1526466}
  for description of this dataset.
  }
\keyword{datasets}
