\name{SMI}
\alias{SMI}
\alias{y}
\docType{data}
\title{Swiss Market Index data}
\usage{
data("SMI")
}
\description{
  See Mullen et al. (forthcoming) for description of this dataset.
}
\references{
  Mullen, K.M., Ardia, D., Gil, D.L, Windover, D., Cline, J. (forthcoming)
  \pkg{DEoptim}: An \R Package for Global Optimization by Differential Evolution.
  \emph{Journal of Statistical Software}, in press.
  URL \url{http://ssrn.com/abstract=1526466}
}
\keyword{datasets}
