\name{SMI}
\alias{SMI}
\alias{y}
\docType{data}
\title{Example Swiss Market Index data}
\usage{
data("SMI")
}
\description{See
  Mullen, K.M., Ardia, D., Gil, D.L, Windover, D., Cline, J. (2009).
  DEoptim: An R Package for Global Optimization by Differential Evolution.
  \url{http://ssrn.com/abstract=1526466}
  for description of this dataset.
  }
\keyword{datasets}
