\name{xrrData}
\alias{xrrData}
\docType{data}
\title{X-ray reflectometry data}
\usage{
data("xrrData")
}
\description{
  See Mullen et al. (2011) for description of this dataset.
}
\references{
  Mullen, K.M, Ardia, D., Gil, D., Windover, D., Cline,
  J. (2011). \emph{\pkg{DEoptim:} An R Package for Global Optimization by
  Differential Evolution}. Journal of Statistical Software, 40(6), 1-26.
  URL \url{http://www.jstatsoft.org/v40/i06/}.
}
\keyword{datasets}
