\name{xrrData}
\alias{xrrData}
\docType{data}
\title{X-ray reflectometry data}
\usage{
data("xrrData")
}
\description{
  See Mullen et al. (2009) for description of this dataset.
}
\references{
  Mullen, K.M., Ardia, D., Gil, D.L, Windover, D., Cline, J. (2009)
  \pkg{DEoptim}: An \R Package for Global Optimization by Differential Evolution.
  URL \url{http://ssrn.com/abstract=1526466}
}
\keyword{datasets}
